requirements.txt
streamlit==1.31.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.35
langchain-groq==0.0.1
huggingface-hub==0.21.3
chromadb==0.3.29
sentence-transformers==2.4.0
unstructured==0.12.4
python-docx==1.1.0
python-dotenv==1.0.1
pysqlite3-binary==0.5.1
